hall work4 and end of draft1
I added lights to add more to really sell the perspective. It still feels a bit weird but maybe that's because I've been looking at it too long. I also made it so the mouse position affects the speed the hallway moves. I closed the screen loop to limit it to the 3 existing screens. I also fixed the infinite  loop bug with screen 0. This will be the draft used for critique.