start of revisions
I started with screen 1. I added it so that the target repels from the mouse. I originally planned for decoys to spawn from the sides but now want to add it so  the decoys spawn in the main one and push out in random directions to confuse the player. I've hit some brick walls so I'll move on to screen 2 and come back.